separated crud and db session functionality
